cvat/requirements.txt
# CVAT Migration Scripts Requirements
Pillow>=9.0.0
requests>=2.28.0
